+++ lib/browser-pool/basic-pool.js
@@ -1,34 +1,61 @@
 'use strict';
-var q = require('q');
+var util = require('util'),
+    Browser = require('../browser'),
+    Pool = require('./pool'),
+    signalHandler = require('../signal-handler'),
+    _ = require('lodash'),
+    Q = require('q');
+
+var activeSessions = {};
+
 /**
  * @constructor
+ * @extends BasicPool
+ * @param {Config} config
+ * @param {Calibrator} calibrator
  */
-function BasicPool() {
+function BasicPool(config, calibrator) {
+    this._config = config;
+    this._calibrator = calibrator;
 }
 
-/**
- * @param {String} id
- * @returns {Promise.<Browser>}
- */
-BasicPool.prototype.getBrowser = function(id) {
+util.inherits(BasicPool, Pool);
+
+Pool.prototype.getBrowser = function(id) {
+    var _this = this,
+        browser = new Browser(this._config.forBrowser(id)),
+        launchPromise = browser.launch(this._calibrator);
+
+    activeSessions[browser.id] = {
+        browser: browser,
+        launchPromise: launchPromise
+    };
+
+    return launchPromise
+        .then(browser.reset.bind(browser))
+        .thenResolve(browser)
+        .fail(function(e) {
+            return _this.freeBrowser(browser)
+                .then(function() {
+                    return Q.reject(e);
+                });
+        });
 };
 
-/**
- * @param {Browser} browser
- * @returns {Promise}
- */
-BasicPool.prototype.freeBrowser = function(browser) {
-    return q.resolve();
+Pool.prototype.freeBrowser = function(browser) {
+    delete activeSessions[browser.id];
+    return browser.quit();
 };
 
-/**
- * @param {String} id
- */
-BasicPool.prototype.finalizeBrowsers = function(id) {
-    return q.resolve();
-};
-
-BasicPool.prototype.cancel = function() {
-};
+signalHandler.on('exit', function() {
+    console.log('Killing browsers...');
+    return _(activeSessions)
+        .map(function(session) {
+            var quit_ = session.browser.quit.bind(session.browser);
+            return Q.when(session.launchPromise, quit_);
+        })
+        .thru(Q.all)
+        .value();
+});
 
 module.exports = BasicPool;
