+++ lib/browser-pool/caching-pool.js
@@ -1,7 +1,7 @@
 'use strict';
 var q = require('q'),
     util = require('util'),
-    BasicPool = require('./basic-pool'),
+    Pool = require('./pool'),
     LimitedUseSet = require('./limited-use-set'),
     log = require('debug')('gemini:pool:caching');
 
@@ -25,7 +25,7 @@ function CachingPool(config, underlyingPool) {
     });
 }
 
-util.inherits(CachingPool, BasicPool);
+util.inherits(CachingPool, Pool);
 
 CachingPool.prototype.getBrowser = function(id) {
     log('request for %s', id);
