+++ lib/browser-pool/index.js
@@ -1,5 +1,5 @@
 'use strict';
-var Pool = require('./pool'),
+var BasicPool = require('./basic-pool'),
     LimitedPool = require('./limited-pool'),
     PerBrowserLimitedPool = require('./per-browser-limited-pool'),
     CachingPool = require('./caching-pool'),
@@ -11,7 +11,7 @@ var Pool = require('./pool'),
  */
 exports.create = function(config) {
     var calibrator = new Calibrator(),
-        pool = new Pool(config, calibrator);
+        pool = new BasicPool(config, calibrator);
 
     pool = new LimitedPool(config.system.parallelLimit, pool);
     pool = new CachingPool(config, pool);
