+++ lib/browser-pool/limited-pool.js
@@ -1,6 +1,6 @@
 'use strict';
 var util = require('util'),
-    BasicPool = require('./basic-pool'),
+    Pool = require('./pool'),
     q = require('q'),
     log = require('debug')('gemini:pool:limited');
 
@@ -17,7 +17,7 @@ function LimitedPool(limit, underlyingPool) {
     this._deferQueue = [];
 }
 
-util.inherits(LimitedPool, BasicPool);
+util.inherits(LimitedPool, Pool);
 
 LimitedPool.prototype.getBrowser = function(id) {
     log('get browser %s (launched %d, limit %d)', id, this._launched, this._limit);
