+++ lib/browser-pool/per-browser-limited-pool.js
@@ -1,7 +1,7 @@
 'use strict';
 var util = require('util'),
     _ = require('lodash'),
-    BasicPool = require('./basic-pool'),
+    Pool = require('./pool'),
     LimitedPool = require('./limited-pool'),
     log = require('debug')('gemini:pool:per-browser-limited');
 
@@ -14,7 +14,7 @@ function PerBrowserLimitedPool(config, underlyingPool) {
     this._browserPools = browserPools;
 }
 
-util.inherits(PerBrowserLimitedPool, BasicPool);
+util.inherits(PerBrowserLimitedPool, Pool);
 
 PerBrowserLimitedPool.prototype.getBrowser = function(id) {
     log('request %s', id);
